+++ Firmware/Board/v3/board.cpp
@@ -58,20 +58,26 @@ OnboardThermistorCurrentLimiter fet_thermistors[AXIS_COUNT] = {
     }
 };
 
+OffboardThermistorCurrentLimiter motor_thermistors[AXIS_COUNT];
+
 Motor motors[AXIS_COUNT] = {
     {
         &htim1, // timer
         TIM_1_8_PERIOD_CLOCKS, // control_deadline
         1.0f / SHUNT_RESISTANCE, // shunt_conductance [S]
         m0_gate_driver, // gate_driver
-        m0_gate_driver // opamp
+        m0_gate_driver, // opamp
+        fet_thermistors[0],
+        motor_thermistors[0]
     },
     {
         &htim8, // timer
         (3 * TIM_1_8_PERIOD_CLOCKS) / 2, // control_deadline
         1.0f / SHUNT_RESISTANCE, // shunt_conductance [S]
         m1_gate_driver, // gate_driver
-        m1_gate_driver // opamp
+        m1_gate_driver, // opamp
+        fet_thermistors[1],
+        motor_thermistors[1]
     }
 };
 
@@ -101,7 +107,6 @@ MechanicalBrake mechanical_brakes[AXIS_COUNT];
 SensorlessEstimator sensorless_estimators[AXIS_COUNT];
 Controller controllers[AXIS_COUNT];
 TrapezoidalTrajectory trap[AXIS_COUNT];
-OffboardThermistorCurrentLimiter motor_thermistors[AXIS_COUNT];
 
 std::array<Axis, AXIS_COUNT> axes{{
     {
@@ -112,8 +117,6 @@ std::array<Axis, AXIS_COUNT> axes{{
         encoders[0], // encoder
         sensorless_estimators[0], // sensorless_estimator
         controllers[0], // controller
-        fet_thermistors[0], // fet_thermistor
-        motor_thermistors[0], // motor_thermistor
         motors[0], // motor
         trap[0], // trap
         endstops[0], endstops[1], // min_endstop, max_endstop
@@ -132,8 +135,6 @@ std::array<Axis, AXIS_COUNT> axes{{
         encoders[1], // encoder
         sensorless_estimators[1], // sensorless_estimator
         controllers[1], // controller
-        fet_thermistors[1], // fet_thermistor
-        motor_thermistors[1], // motor_thermistor
         motors[1], // motor
         trap[1], // trap
         endstops[2], endstops[3], // min_endstop, max_endstop